move out teams from microplanning api
