+++ hat/assets/js/apps/Iaso/domains/assignments/hooks/requests/useGetTeams.ts
@@ -9,7 +9,7 @@ import { Teams, DropdownTeamsOptions, Team } from '../../types/team';
 import { getColor } from '../../constants/colors';
 
 const getTeams = async (ancestor: number): Promise<Teams> => {
-    const url = makeUrlWithParams('/api/microplanning/teams/', {
+    const url = makeUrlWithParams('/api/teams/', {
         ancestor: `${ancestor}`,
     });
     return getRequest(url) as Promise<Teams>;
