+++ iaso/api/teams/__init__.py
@@ -1,0 +1,4 @@
+from .views import TeamViewSet
+
+__all__ = ["TeamViewSet"]
+
